+++ src/main/java/org/ow2/authzforce/pap/dao/flatfile/FlatFileBasedDomainDAO.java
@@ -22,7 +22,7 @@ import java.io.IOException;
 import org.ow2.authzforce.core.pap.api.dao.DomainDAO;
 import org.ow2.authzforce.core.pap.api.dao.PolicyDAOClient;
 import org.ow2.authzforce.core.pap.api.dao.PolicyVersionDAOClient;
-import org.ow2.authzforce.pap.dao.file.xmlns.DomainProperties;
+import org.ow2.authzforce.pap.dao.flatfile.xmlns.DomainProperties;
 
 interface FlatFileBasedDomainDAO<VERSION_DAO_CLIENT extends PolicyVersionDAOClient, POLICY_DAO_CLIENT extends PolicyDAOClient>
 		extends DomainDAO<VERSION_DAO_CLIENT, POLICY_DAO_CLIENT>, Closeable
@@ -47,8 +47,7 @@ interface FlatFileBasedDomainDAO<VERSION_DAO_CLIENT extends PolicyVersionDAOClie
 	 * @return up-to-date domain properties (after sync)
 	 * 
 	 * @throws IllegalArgumentException
-	 *             Invalid PDP configuration files (e.g. policies or PDP
-	 *             configuration)
+	 *             Invalid PDP configuration files (e.g. policies or PDP configuration)
 	 * @throws IOException
 	 *             Problem getting PDP configuration files from repository
 	 */